+++ ci/install.sh
@@ -1,9 +1,7 @@
 set -euxo pipefail
 
 main() {
-    if [ $TARGET != x86_64-unknown-linux-gnu ]; then
-        rustup target add $TARGET
-    fi
+    rustup target add $TARGET
 }
 
 main