+++ CHANGELOG.md
@@ -5,6 +5,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [v0.3.2] - 2019-11-26
+
+### Added
+
+- `Aligned<_, [T]>` now implements the `Index<RangeTo<usize>>` trait; slicing
+  this value to end returns an `Aligned<_, [T]>` slice.
+
 ## [v0.3.1] - 2018-11-07
 
 ### Changed
@@ -48,7 +55,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Initial release
 
-[Unreleased]: https://github.com/japaric/aligned/compare/v0.3.0...HEAD
+[Unreleased]: https://github.com/japaric/aligned/compare/v0.3.2...HEAD
+[v0.3.2]: https://github.com/japaric/aligned/compare/v0.3.1...v0.3.2
+[v0.3.1]: https://github.com/japaric/aligned/compare/v0.3.0...v0.3.1
 [v0.3.0]: https://github.com/japaric/aligned/compare/v0.2.0...v0.3.0
 [v0.2.0]: https://github.com/japaric/aligned/compare/v0.1.2...v0.2.0
 [v0.1.2]: https://github.com/japaric/aligned/compare/v0.1.1...v0.1.2